Separadas funções em arquivos.
